+++ pylabrobot/resources/porvair/plates.py
@@ -2,6 +2,7 @@
 
 # pylint: disable=invalid-name
 
+from typing import Optional
 from pylabrobot.resources.plate import Lid, Plate
 from pylabrobot.resources.well import Well, WellBottomType, CrossSectionType
 from pylabrobot.resources.utils import create_ordered_items_2d
@@ -92,3 +93,55 @@ def Porvair_6_reservoir_47ml_Vb_L(name: str, with_lid: bool = False) -> Plate:
 #: Porvair_6_reservoir_47ml_Vb_P
 def Porvair_6_reservoir_47ml_Vb_P(name: str, with_lid: bool = False) -> Plate:
   return Porvair_6_reservoir_47ml_Vb(name=name, with_lid=with_lid).rotated(z=90)
+
+
+def Porvair_24_wellplate_Vb(name: str, lid: Optional[Lid] = None) -> Plate:
+  """
+  Porvair cat. no.: 390108
+  - Material: Polypropylene
+  - Tissue culture treated: No
+  """
+  WELL_SIZE_X = 8.0
+  WELL_SIZE_Y = 35.0
+  WELL_HEIGHT_OF_PYRAMID = 11.65
+  WELL_HEIGHT_OF_CUBE = 1.65
+
+  well_kwargs = {
+    "size_x": 8.0,
+    "size_y": 35.0,
+    "size_z": 13.30,
+    # reality: multifaceted pyramid, v bottom pyramid is a good approximation
+    "bottom_type": WellBottomType.V,
+    "compute_volume_from_height":
+      lambda liquid_height: calculate_liquid_volume_container_2segments_square_vbottom(
+      WELL_SIZE_X, WELL_SIZE_Y, WELL_HEIGHT_OF_PYRAMID, WELL_HEIGHT_OF_CUBE, liquid_height
+    ),
+    "compute_height_from_volume": lambda liquid_volume: (
+      calculate_liquid_height_in_container_2segments_square_vbottom(
+        WELL_SIZE_X, WELL_SIZE_Y, WELL_HEIGHT_OF_PYRAMID, WELL_HEIGHT_OF_CUBE, liquid_volume
+      )
+    ),
+    "material_z_thickness": 1.0,  # measured
+    "cross_section_type": CrossSectionType.RECTANGLE,
+    "max_volume": 3500,  # according to porvair spec
+  }
+
+  return Plate(
+    name=name,
+    size_x=127.5,  # measured
+    size_y=85.3,  # measured
+    size_z=19.2,  # measured
+    lid=lid,
+    model=Porvair_24_wellplate_Vb.__name__,
+    ordered_items=create_ordered_items_2d(
+      Well,
+      num_items_x=12,
+      num_items_y=2,
+      dx=9.90,  # measured
+      dy=6.85,  # measured
+      dz=5.8,  # measured and calibrated manually
+      item_dx=9.05,
+      item_dy=36,
+      **well_kwargs,
+    ),
+  )